FEAT: Split provider gtests and add more tests
